+++ src/main/java/com/example/SteDziPanki/model/Continent.java
@@ -9,7 +9,7 @@ public class Continent {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @Column(length = 100)
     private String name;
 
@@ -25,17 +25,17 @@ public class Continent {
     public Continent() {
     }
 
-    public Continent(Integer id, String name, Set<Country> countries) {
+    public Continent(Long id, String name, Set<Country> countries) {
         this.id = id;
         this.name = name;
         this.countries = countries;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
